Changed num_bins from float to int
Changed CG_ENG_MPIC_WAKE_CPM from float to int

Changed CG_ENG_MPIC_WAKE_CPM from float to int

Changed LAMP_TIME from float to int

convert CG_ENG_MPIC_WAKE_CPM parameter from float to int

+++ mi/dataset/parser/adcpt_m_log9.py
@@ -87,7 +87,7 @@ LOG9_DATA_MAP = [
     ('h_1_10', 18, float),
     ('t_1_10', 19, float),
     ('d_mean', 20, float),
-    ('num_bins', 21, float),
+    ('num_bins', 21, int),
     ('depth_level_magnitude', 22, list),
     ('depth_level_direction', 23, list),]
